Use a custom Vertx worker pool with size matching database pool
This avoid blocking workers (at least in the “normal” case when all
connections in the pool are connected), which vastly improves
concurrency. For example you can create two pools, one for fast queries
and one for slow, and even if you saturate the slow pool the fast one
will not block. This required upgrading to 3.3.2 of Vert.x.